+++ packages/cinterop/src/darwin/kotlin/io/realm/kotlin/internal/interop/RealmInterop.kt
@@ -19,6 +19,9 @@
 package io.realm.kotlin.internal.interop
 
 import io.realm.kotlin.internal.interop.Constants.ENCRYPTION_KEY_LENGTH
+import io.realm.kotlin.internal.interop.RealmInterop.asByteArray
+import io.realm.kotlin.internal.interop.RealmInterop.asTimestamp
+import io.realm.kotlin.internal.interop.RealmInterop.safeKString
 import io.realm.kotlin.internal.interop.sync.AppError
 import io.realm.kotlin.internal.interop.sync.AppErrorCategory
 import io.realm.kotlin.internal.interop.sync.AuthProvider
@@ -98,6 +101,7 @@ import realm_wrapper.realm_link_t
 import realm_wrapper.realm_list_t
 import realm_wrapper.realm_object_t
 import realm_wrapper.realm_property_info_t
+import realm_wrapper.realm_query_arg_t
 import realm_wrapper.realm_release
 import realm_wrapper.realm_scheduler_notify_func_t
 import realm_wrapper.realm_scheduler_t
@@ -177,8 +181,7 @@ fun realm_string_t.set(memScope: MemScope, s: String): realm_string_t {
 
 @Suppress("LongMethod", "ComplexMethod")
 fun realm_value_t.set(memScope: MemScope, realmValue: RealmValue): realm_value_t {
-    val value = realmValue.value
-    when (value) {
+    when (val value = realmValue.value) {
         null -> {
             type = realm_value_type.RLM_TYPE_NULL
         }
@@ -574,9 +577,19 @@ actual object RealmInterop {
         }
     }
 
-    actual fun realm_convert_with_config(realm: RealmPointer, config: RealmConfigurationPointer) {
+    actual fun realm_convert_with_config(
+        realm: RealmPointer,
+        config: RealmConfigurationPointer,
+        mergeWithExisting: Boolean
+    ) {
         memScoped {
-            checkedBooleanResult(realm_wrapper.realm_convert_with_config(realm.cptr(), config.cptr()))
+            checkedBooleanResult(
+                realm_wrapper.realm_convert_with_config(
+                    realm.cptr(),
+                    config.cptr(),
+                    mergeWithExisting
+                )
+            )
         }
     }
 
@@ -1122,19 +1135,13 @@ actual object RealmInterop {
     ): RealmQueryPointer {
         memScoped {
             val count = args.size
-            val cArgs = allocArray<realm_value_t>(count)
-            args.mapIndexed { i, arg ->
-                cArgs[i].apply {
-                    set(memScope, arg)
-                }
-            }
             return CPointerWrapper(
                 realm_wrapper.realm_query_parse(
                     realm.cptr(),
                     classKey.key.toUInt(),
                     query,
                     count.toULong(),
-                    cArgs
+                    args.toQueryArgs(this)
                 )
             )
         }
@@ -1147,18 +1154,12 @@ actual object RealmInterop {
     ): RealmQueryPointer {
         memScoped {
             val count = args.size
-            val cArgs = allocArray<realm_value_t>(count)
-            args.mapIndexed { i, arg ->
-                cArgs[i].apply {
-                    set(memScope, arg)
-                }
-            }
             return CPointerWrapper(
                 realm_wrapper.realm_query_parse_for_results(
                     results.cptr(),
                     query,
                     count.toULong(),
-                    cArgs
+                    args.toQueryArgs(this)
                 )
             )
         }
@@ -1204,18 +1205,12 @@ actual object RealmInterop {
     ): RealmQueryPointer {
         memScoped {
             val count = args.size
-            val cArgs = allocArray<realm_value_t>(count)
-            args.mapIndexed { i, arg ->
-                cArgs[i].apply {
-                    set(memScope, arg)
-                }
-            }
             return CPointerWrapper(
                 realm_wrapper.realm_query_append_query(
                     query.cptr(),
                     filter,
                     count.toULong(),
-                    cArgs
+                    args.toQueryArgs(this)
                 )
             )
         }
@@ -1861,7 +1856,15 @@ actual object RealmInterop {
             staticCFunction { userData, beforeRealm ->
                 val beforeCallback = safeUserData<SyncBeforeClientResetHandler>(userData)
                 val beforeDb = CPointerWrapper<FrozenRealmT>(beforeRealm, false)
-                beforeCallback.onBeforeReset(beforeDb)
+
+                // Check if exceptions have been thrown, return true if all went as it should
+                try {
+                    beforeCallback.onBeforeReset(beforeDb)
+                    true
+                } catch (e: Throwable) {
+                    println(e.message)
+                    false
+                }
             },
             StableRef.create(beforeHandler.freeze()).asCPointer(),
             staticCFunction { userdata ->
@@ -1879,8 +1882,19 @@ actual object RealmInterop {
             staticCFunction { userData, beforeRealm, afterRealm, didRecover ->
                 val afterCallback = safeUserData<SyncAfterClientResetHandler>(userData)
                 val beforeDb = CPointerWrapper<FrozenRealmT>(beforeRealm, false)
-                val afterDb = CPointerWrapper<LiveRealmT>(afterRealm, false)
-                afterCallback.onAfterReset(beforeDb, afterDb, didRecover)
+
+                // afterRealm is wrapped inside a ThreadSafeReference so the pointer needs to be resolved
+                val afterRealmPtr = realm_wrapper.realm_from_thread_safe_reference(afterRealm, null)
+                val afterDb = CPointerWrapper<LiveRealmT>(afterRealmPtr, false)
+
+                // Check if exceptions have been thrown, return true if all went as it should
+                try {
+                    afterCallback.onAfterReset(beforeDb, afterDb, didRecover)
+                    true
+                } catch (e: Throwable) {
+                    println(e.message)
+                    false
+                }
             },
             StableRef.create(afterHandler.freeze()).asCPointer(),
             staticCFunction { userdata ->
@@ -2016,8 +2030,8 @@ actual object RealmInterop {
         realm_wrapper.realm_app_config_set_base_url(appConfig.cptr(), baseUrl)
     }
 
-    actual fun realm_app_credentials_new_anonymous(): RealmCredentialsPointer {
-        return CPointerWrapper(realm_wrapper.realm_app_credentials_new_anonymous())
+    actual fun realm_app_credentials_new_anonymous(reuseExisting: Boolean): RealmCredentialsPointer {
+        return CPointerWrapper(realm_wrapper.realm_app_credentials_new_anonymous(reuseExisting))
     }
 
     actual fun realm_app_credentials_new_email_password(
@@ -2425,6 +2439,35 @@ actual object RealmInterop {
         mutableSubscriptionSet: RealmMutableSubscriptionSetPointer
     ): RealmSubscriptionSetPointer {
         return CPointerWrapper(realm_wrapper.realm_sync_subscription_set_commit(mutableSubscriptionSet.cptr()))
+    }
+
+    /**
+     * C-API functions for queries receive a pointer to one or more 'realm_query_arg_t' query
+     * arguments. In turn, said arguments contain individual values or lists of values (in
+     * combination with the 'is_list' flag) in order to support predicates like
+     *
+     * "fruit IN {'apple', 'orange'}"
+     *
+     * which is a statement equivalent to
+     *
+     * "fruit == 'apple' || fruit == 'orange'"
+     *
+     * See https://github.com/realm/realm-core/issues/4266 for more info.
+     */
+    private fun Array<RealmValue>.toQueryArgs(memScope: MemScope): CPointer<realm_query_arg_t> {
+        with(memScope) {
+            val cArgs = allocArray<realm_query_arg_t>(this@toQueryArgs.size)
+            this@toQueryArgs.mapIndexed { i, arg ->
+                val value = alloc<realm_value_t>()
+                    .set(this, arg)
+                cArgs[i].apply {
+                    this.nb_args = 1.toULong()
+                    this.is_list = false
+                    this.arg = value.ptr
+                }
+            }
+            return cArgs
+        }
     }
 
     private fun <T : CapiT> nativePointerOrNull(ptr: CPointer<*>?, managed: Boolean = true): NativePointer<T>? {