+++ packages/cinterop/src/jvm/kotlin/io/realm/kotlin/internal/interop/sync/SyncSessionResyncMode.kt
@@ -21,7 +21,9 @@ import io.realm.kotlin.internal.interop.realm_sync_session_resync_mode_e
 
 actual enum class SyncSessionResyncMode(override val nativeValue: Int) : NativeEnumerated {
     RLM_SYNC_SESSION_RESYNC_MODE_MANUAL(realm_sync_session_resync_mode_e.RLM_SYNC_SESSION_RESYNC_MODE_MANUAL),
-    RLM_SYNC_SESSION_RESYNC_MODE_DISCARD_LOCAL(realm_sync_session_resync_mode_e.RLM_SYNC_SESSION_RESYNC_MODE_DISCARD_LOCAL);
+    RLM_SYNC_SESSION_RESYNC_MODE_DISCARD_LOCAL(realm_sync_session_resync_mode_e.RLM_SYNC_SESSION_RESYNC_MODE_DISCARD_LOCAL),
+    RLM_SYNC_SESSION_RESYNC_MODE_RECOVER(realm_sync_session_resync_mode_e.RLM_SYNC_SESSION_RESYNC_MODE_RECOVER),
+    RLM_SYNC_SESSION_RESYNC_MODE_RECOVER_OR_DISCARD(realm_sync_session_resync_mode_e.RLM_SYNC_SESSION_RESYNC_MODE_RECOVER_OR_DISCARD);
 
     actual companion object {
         actual fun fromInt(nativeValue: Int): SyncSessionResyncMode {